Responsive draft
Responsive first draft which allow to display the website at lower desktop & mobile.
