+++ src/app/_mixins.scss
@@ -1,0 +1,11 @@
+@mixin small-desktop {
+  @media (max-width: 1200px) {
+    @content;
+  }
+}
+
+@mixin mobile {
+  @media (max-width: 700px) {
+    @content;
+  }
+}
