+++ src/app/components/Footer/Footer.module.scss
@@ -1,19 +1,40 @@
+@use "../../mixins" as *;
+
 .footer {
   z-index: 3;
   display: flex;
   flex-direction: row;
   align-items: center;
   justify-content: space-between;
+
   position: fixed;
   bottom: 0;
   width: 100%;
   padding: 30px;
   background-color: var(--bg-body-color);
   border-top: 1px solid var(--color-3);
+
+  @include small-desktop {
+    flex-direction: column;
+    align-items: flex-start;
+    gap: var(--gap-footer);
+  }
 }
 
 .group {
   display: flex;
   flex-direction: row;
-  gap: 40px;
+  gap: var(--gap-footer);
+
+  a {
+    font-size: var (--font-size-very-big);
+  }
+
+  @include small-desktop {
+    width: 100%;
+    justify-content: center;
+    a {
+      font-size: var(--font-size-very-small);
+    }
+  }
 }
