+++ src/app/components/Header/Header.module.scss
@@ -1,3 +1,5 @@
+@use "../../mixins" as *;
+
 .header {
   padding-bottom: var(--header-padding-bottom);
 }
@@ -7,4 +9,5 @@
   position: absolute;
   justify-content: space-between;
   width: 100%;
+  @include small-desktop { display: none}
 }
