+++ src/app/components/Header/Header.tsx
@@ -3,7 +3,7 @@
 import styles from "./Header.module.scss"
 import { CodeBlock } from "../CodeBlock/CodeBlock";
 import { codeSnipetsHeader, codeSnipetsHeader2 } from "../../data/code";
-import { useEffect, useRef } from "react";
+import { useEffect, useRef, useState } from "react";
 import { motion } from "framer-motion";
 import { usePathname } from 'next/navigation';
 
@@ -11,12 +11,12 @@ import { usePathname } from 'next/navigation';
 export function Header(): JSX.Element {
 
 	const pathname: string = usePathname();
-
 	/*Using useRef and ++ to regenerate <Header/> and play the animation each time the pathname changes. **/
 	const ref = useRef<number>(0);
 	useEffect(() => {
 		ref.current++;
 	}, [pathname])
+
 
 	let codeSnipetstoShow = codeSnipetsHeader2;
 	if (pathname.length < 2) { codeSnipetstoShow = codeSnipetsHeader; }
@@ -36,5 +36,6 @@ export function Header(): JSX.Element {
 			</motion.div>
 		</header>
 	);
+
 
 }
