+++ src/app/components/MugCoffee/MugCoffee.module.scss
@@ -1,4 +1,5 @@
 .mugcoffee {
+  z-index: 1;
   filter: drop-shadow(-30px 15px 11px rgba(0, 0, 0, 0.7));
 
   display: flex;
