+++ src/app/components/Tag/Tag.scss
@@ -1,4 +1,10 @@
 .tag {
+  font-size: var(--font-size-very-small);
+  border-radius: 50px;
+  color: var(--color-1);
+  padding: 5px;
+  transition: 0.15s all ease;
+
   $tagcolor: (
     react: #003d66,
     typescript: #2485e2,
@@ -16,11 +22,7 @@
   @mixin generateColors($tagcolor) {
     @each $name, $color in $tagcolor {
       &--#{$name} {
-        font-size: var(--font-size-very-small);
-        border-radius: 50px;
-        color: var(--color-1);
-        padding: 5px;
-        transition: 0.15s all ease;
+        @extend .tag;
         border: 2px solid $color;
         &:hover {
           color: white;
