+++ src/app/page.module.scss
@@ -1,3 +1,4 @@
+@use "./mixins" as *;
 .home {
   display: flex;
   flex-direction: column;
